update teardown to only delete newly created test tables
- refactor getConfig and deleteTables to new utils
- add getRelevantTables util
- refactor setup to use new utils
- update setup and teardown to only attempt deletion of tables created by setup
