+++ src/teardown.ts
@@ -1,5 +1,8 @@
 import DynamoDbLocal from 'dynamodb-local';
 import type {JestArgs} from './types';
+import deleteTables from './utils/delete-tables';
+import getConfig from './utils/get-config';
+import getRelevantTables from './utils/get-relevant-tables';
 
 const debug = require('debug')('jest-dynamodb');
 
@@ -15,10 +18,12 @@ export default async function (jestArgs: JestArgs) {
     }
   } else {
     const dynamoDB = global.__DYNAMODB_CLIENT__;
+    const {tables: targetTables} = await getConfig(debug);
+
     const {TableNames: tableNames} = await dynamoDB.listTables({});
 
     if (tableNames?.length) {
-      await Promise.all(tableNames.map(tableName => dynamoDB.deleteTable({TableName: tableName})));
+      await deleteTables(dynamoDB, getRelevantTables(tableNames, targetTables));
     }
   }
-};
+}
