+++ src/utils/delete-tables.ts
@@ -1,0 +1,5 @@
+import type {DynamoDB} from '@aws-sdk/client-dynamodb';
+
+export default function deleteTables(dynamoDB: DynamoDB, tableNames: string[]) {
+  return Promise.all(tableNames.map(tableName => dynamoDB.deleteTable({TableName: tableName})));
+}
