+++ src/utils/get-relevant-tables.ts
@@ -1,0 +1,10 @@
+import type {CreateTableCommandInput} from '@aws-sdk/client-dynamodb';
+
+export default function getRelevantTables(
+  dbTables: string[],
+  configTables: CreateTableCommandInput[]
+) {
+  const configTableNames = configTables.map(configTable => configTable.TableName);
+
+  return dbTables.filter(dbTable => configTableNames.includes(dbTable));
+}
